pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai-tools>=1.6.0",
    "crewai[google-genai]>=1.6.0",
    "faiss-cpu>=1.13.0",
    "fastapi[standard]>=0.122.0",
    "googlemaps>=4.10.0",
    "hf-xet>=1.2.0",
    "huggingface>=0.0.1",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-vertexai>=3.1.0",
    "langchain-groq>=1.1.0",
    "langchain-huggingface>=1.1.0",
    "langchain-litellm>=0.3.2",
    "langgraph>=1.0.4",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "ser>=0.1.0",
    "setuptools>=80.9.0",
    "streamlit>=1.51.0",
]
